access and refresh cookies were added

+++ backend/crud/views.py
@@ -5,9 +5,12 @@ from rest_framework.response import Response
 from .models import Blog
 from .serializers import BlogSerializer
 
+
 # I`m going to create views with @api_view for understanding the background logic in the code.
 
 @api_view(['GET', 'POST'])
+
+
 def blog_list(request):
 
     if request.method == 'GET':
@@ -16,18 +19,29 @@ def blog_list(request):
         return Response(serializer.data)
     
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            return Response({'detail': 'Authentication required'}, status=401)
+
         serializer = BlogSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(status=201)
         return Response(serializer.errors,status=404)
     
-@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+
 def blog_detail(request, pk):
+
     try:
         blog = Blog.objects.get(pk=pk)
     except Blog.DoesNotExist:
         return Response(status=404)
+    
+    def check_owner_permission(request, blog):
+        if not request.user.is_authenticated:
+            return Response ({'detail': 'Authorization required'}, status=401)
+        if request.user != blog.owner:
+            return Response({'detail': 'Not Allowed'}, status= 403)
+        
     
     if request.method == 'GET':
         serializer = BlogSerializer(blog)
@@ -35,6 +49,10 @@ def blog_detail(request, pk):
   
     
     elif request.method == 'PUT':
+        permission_check = check_owner_permission()
+        if permission_check:
+            return permission_check
+        
         serializer = BlogSerializer(blog, data = request.data)
         if serializer.is_valid():
             serializer.save()
@@ -43,6 +61,10 @@ def blog_detail(request, pk):
     
     
     elif request.method == 'PATCH':
+        permission_check = check_owner_permission()
+        if permission_check:
+            return permission_check
+      
         serializer = BlogSerializer(blog, data = request.data, partial = True)
         if serializer.is_valid():
             serializer.save()
@@ -50,5 +72,9 @@ def blog_detail(request, pk):
         return Response(status=400)
 
     elif request.method == 'DELETE':
+        permission_check = check_owner_permission()
+        if permission_check:
+            return permission_check
+       
         blog.delete()
         return Response(status=204)
